+++ cvat-ui/src/components/annotation-page/attribute-annotation-workspace/attribute-annotation-sidebar/attribute-annotation-sidebar.tsx
@@ -5,23 +5,25 @@
 import React, { useState, useEffect } from 'react';
 import { GlobalHotKeys, ExtendedKeyMapOptions } from 'react-hotkeys';
 import { connect } from 'react-redux';
-import { Action } from 'redux';
-import { ThunkDispatch } from 'redux-thunk';
 import Layout, { SiderProps } from 'antd/lib/layout';
 import { SelectValue } from 'antd/lib/select';
-import { CheckboxChangeEvent } from 'antd/lib/checkbox';
 import { Row, Col } from 'antd/lib/grid';
 import Text from 'antd/lib/typography/Text';
 import Icon from 'antd/lib/icon';
 
+import { ThunkDispatch } from 'utils/redux';
 import { Canvas } from 'cvat-canvas-wrapper';
 import { LogType } from 'cvat-logger';
 import {
     activateObject as activateObjectAction,
     updateAnnotationsAsync,
+    changeFrameAsync,
 } from 'actions/annotation-actions';
-import { CombinedState } from 'reducers/interfaces';
+import { CombinedState, ObjectType } from 'reducers/interfaces';
 import AnnotationsFiltersInput from 'components/annotation-page/annotations-filters-input';
+import AppearanceBlock from 'components/annotation-page/appearance-block';
+import ObjectButtonsContainer from 'containers/annotation-page/standard-workspace/objects-side-bar/object-buttons';
+
 import ObjectSwitcher from './object-switcher';
 import AttributeSwitcher from './attribute-switcher';
 import ObjectBasicsEditor from './object-basics-edtior';
@@ -43,6 +45,7 @@ interface StateToProps {
 interface DispatchToProps {
     activateObject(clientID: number | null, attrID: number | null): void;
     updateAnnotations(statesToUpdate: any[]): void;
+    changeFrame(frame: number): void;
 }
 
 interface LabelAttrMap {
@@ -85,13 +88,16 @@ function mapStateToProps(state: CombinedState): StateToProps {
     };
 }
 
-function mapDispatchToProps(dispatch: ThunkDispatch<CombinedState, {}, Action>): DispatchToProps {
+function mapDispatchToProps(dispatch: ThunkDispatch): DispatchToProps {
     return {
         activateObject(clientID: number, attrID: number): void {
             dispatch(activateObjectAction(clientID, attrID));
         },
         updateAnnotations(states): void {
             dispatch(updateAnnotationsAsync(states));
+        },
+        changeFrame(frame: number): void {
+            dispatch(changeFrameAsync(frame));
         },
     };
 }
@@ -104,6 +110,7 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
         activatedAttributeID,
         jobInstance,
         updateAnnotations,
+        changeFrame,
         activateObject,
         keyMap,
         normalizedKeyMap,
@@ -111,6 +118,7 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
         canvasIsReady,
     } = props;
 
+    const filteredStates = states.filter((state) => !state.outside && !state.hidden);
     const [labelAttrMap, setLabelAttrMap] = useState(
         labels.reduce((acc, label): LabelAttrMap => {
             acc[label.id] = label.attributes.length ? label.attributes[0] : null;
@@ -133,10 +141,11 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
         setSidebarCollapsed(!sidebarCollapsed);
     };
 
-    const [activeObjectState] = activatedStateID === null
-        ? [null] : states.filter((objectState: any): boolean => (
-            objectState.clientID === activatedStateID
-        ));
+    const indexes = filteredStates.map((state) => state.clientID);
+    const activatedIndex = indexes.indexOf(activatedStateID);
+    const activeObjectState = activatedStateID === null || activatedIndex === -1
+        ? null : filteredStates[activatedIndex];
+
     const activeAttribute = activeObjectState
         ? labelAttrMap[activeObjectState.label.id]
         : null;
@@ -147,24 +156,24 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
             if (attribute && attribute.id !== activatedAttributeID) {
                 activateObject(activatedStateID, attribute ? attribute.id : null);
             }
-        } else if (states.length) {
-            const attribute = labelAttrMap[states[0].label.id];
-            activateObject(states[0].clientID, attribute ? attribute.id : null);
+        } else if (filteredStates.length) {
+            const attribute = labelAttrMap[filteredStates[0].label.id];
+            activateObject(filteredStates[0].clientID, attribute ? attribute.id : null);
         }
     }
 
     const nextObject = (step: number): void => {
-        if (states.length) {
-            const index = states.indexOf(activeObjectState);
+        if (filteredStates.length) {
+            const index = filteredStates.indexOf(activeObjectState);
             let nextIndex = index + step;
-            if (nextIndex > states.length - 1) {
+            if (nextIndex > filteredStates.length - 1) {
                 nextIndex = 0;
             } else if (nextIndex < 0) {
-                nextIndex = states.length - 1;
+                nextIndex = filteredStates.length - 1;
             }
             if (nextIndex !== index) {
-                const attribute = labelAttrMap[states[nextIndex].label.id];
-                activateObject(states[nextIndex].clientID, attribute ? attribute.id : null);
+                const attribute = labelAttrMap[filteredStates[nextIndex].label.id];
+                activateObject(filteredStates[nextIndex].clientID, attribute ? attribute.id : null);
             }
         }
     };
@@ -207,41 +216,73 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
         collapsed: sidebarCollapsed,
     };
 
+    const preventDefault = (event: KeyboardEvent | undefined): void => {
+        if (event) {
+            event.preventDefault();
+        }
+    };
+
     const subKeyMap = {
         NEXT_ATTRIBUTE: keyMap.NEXT_ATTRIBUTE,
         PREVIOUS_ATTRIBUTE: keyMap.PREVIOUS_ATTRIBUTE,
         NEXT_OBJECT: keyMap.NEXT_OBJECT,
         PREVIOUS_OBJECT: keyMap.PREVIOUS_OBJECT,
+        SWITCH_LOCK: keyMap.SWITCH_LOCK,
+        SWITCH_OCCLUDED: keyMap.SWITCH_OCCLUDED,
+        NEXT_KEY_FRAME: keyMap.NEXT_KEY_FRAME,
+        PREV_KEY_FRAME: keyMap.PREV_KEY_FRAME,
     };
 
     const handlers = {
         NEXT_ATTRIBUTE: (event: KeyboardEvent | undefined) => {
-            if (event) {
-                event.preventDefault();
-            }
-
+            preventDefault(event);
             nextAttribute(1);
         },
         PREVIOUS_ATTRIBUTE: (event: KeyboardEvent | undefined) => {
-            if (event) {
-                event.preventDefault();
-            }
-
+            preventDefault(event);
             nextAttribute(-1);
         },
         NEXT_OBJECT: (event: KeyboardEvent | undefined) => {
-            if (event) {
-                event.preventDefault();
-            }
-
+            preventDefault(event);
             nextObject(1);
         },
         PREVIOUS_OBJECT: (event: KeyboardEvent | undefined) => {
-            if (event) {
-                event.preventDefault();
-            }
-
+            preventDefault(event);
             nextObject(-1);
+        },
+        SWITCH_LOCK: (event: KeyboardEvent | undefined) => {
+            preventDefault(event);
+            if (activeObjectState) {
+                activeObjectState.lock = !activeObjectState.lock;
+                updateAnnotations([activeObjectState]);
+            }
+        },
+        SWITCH_OCCLUDED: (event: KeyboardEvent | undefined) => {
+            preventDefault(event);
+            if (activeObjectState && activeObjectState.objectType !== ObjectType.TAG) {
+                activeObjectState.occluded = !activeObjectState.occluded;
+                updateAnnotations([activeObjectState]);
+            }
+        },
+        NEXT_KEY_FRAME: (event: KeyboardEvent | undefined) => {
+            preventDefault(event);
+            if (activeObjectState && activeObjectState.objectType === ObjectType.TRACK) {
+                const frame = typeof (activeObjectState.keyframes.next) === 'number'
+                    ? activeObjectState.keyframes.next : null;
+                if (frame !== null && canvasInstance.isAbleToChangeFrame()) {
+                    changeFrame(frame);
+                }
+            }
+        },
+        PREV_KEY_FRAME: (event: KeyboardEvent | undefined) => {
+            preventDefault(event);
+            if (activeObjectState && activeObjectState.objectType === ObjectType.TRACK) {
+                const frame = typeof (activeObjectState.keyframes.prev) === 'number'
+                    ? activeObjectState.keyframes.prev : null;
+                if (frame !== null && canvasInstance.isAbleToChangeFrame()) {
+                    changeFrame(frame);
+                }
+            }
         },
     };
 
@@ -268,15 +309,14 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
                     currentLabel={activeObjectState.label.name}
                     clientID={activeObjectState.clientID}
                     occluded={activeObjectState.occluded}
-                    objectsCount={states.length}
-                    currentIndex={states.indexOf(activeObjectState)}
+                    objectsCount={filteredStates.length}
+                    currentIndex={filteredStates.indexOf(activeObjectState)}
                     normalizedKeyMap={normalizedKeyMap}
                     nextObject={nextObject}
                 />
                 <ObjectBasicsEditor
                     currentLabel={activeObjectState.label.name}
                     labels={labels}
-                    occluded={activeObjectState.occluded}
                     changeLabel={(value: SelectValue): void => {
                         const labelName = value as string;
                         const [newLabel] = labels
@@ -284,10 +324,12 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
                         activeObjectState.label = newLabel;
                         updateAnnotations([activeObjectState]);
                     }}
-                    setOccluded={(event: CheckboxChangeEvent): void => {
-                        activeObjectState.occluded = event.target.checked;
-                        updateAnnotations([activeObjectState]);
-                    }}
+                />
+                <ObjectButtonsContainer
+                    clientID={activeObjectState.clientID}
+                    outsideDisabled
+                    hiddenDisabled
+                    keyframeDisabled
                 />
                 {
                     activeAttribute
@@ -302,6 +344,7 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
                                     nextAttribute={nextAttribute}
                                 />
                                 <AttributeEditor
+                                    clientID={activeObjectState.clientID}
                                     attribute={activeAttribute}
                                     currentValue={activeObjectState.attributes[activeAttribute.id]}
                                     onChange={(value: string) => {
@@ -326,12 +369,29 @@ function AttributeAnnotationSidebar(props: StateToProps & DispatchToProps): JSX.
                             </div>
                         )
                 }
+
+                { !sidebarCollapsed && <AppearanceBlock /> }
             </Layout.Sider>
         );
     }
 
     return (
         <Layout.Sider {...siderProps}>
+            {/* eslint-disable-next-line */}
+            <span
+                className={`cvat-objects-sidebar-sider
+                    ant-layout-sider-zero-width-trigger
+                    ant-layout-sider-zero-width-trigger-left`}
+                onClick={collapse}
+            >
+                {sidebarCollapsed ? <Icon type='menu-fold' title='Show' />
+                    : <Icon type='menu-unfold' title='Hide' />}
+            </span>
+            <Row className='cvat-objects-sidebar-filter-input'>
+                <Col>
+                    <AnnotationsFiltersInput />
+                </Col>
+            </Row>
             <div className='attribute-annotations-sidebar-not-found-wrapper'>
                 <Text strong>No objects found</Text>
             </div>